Rename strings.cpp and strings.h to appletstrings.cpp and appletstrings.h
