+++ appletstrings.cpp
@@ -22,7 +22,7 @@
  *
  * END_COMMON_COPYRIGHT_HEADER */
 
-#include "strings.h"
+#include "appletstrings.h"
 
 #include <QMap>
 #include <QObject>
