+++ appletstrings.h
